Feat: seperated component folders, css changes, and renamed the file for the Profile Button
