+++ frontend/src/components/Preferences.jsx
@@ -1,0 +1,16 @@
+
+
+export default function Preferences() {
+
+
+    return (
+
+    <div>
+      <h2 className="text-lg font-semibold">Preferences</h2>
+      <p className="text-sm text-icon dark:text-icon-dark">
+        More to come
+      </p>
+    </div>
+
+  );
+}
