+++ frontend/src/components/Settings.jsx
@@ -1,10 +1,14 @@
-import { useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { IoCloseSharp } from "react-icons/io5";
 import { HiOutlineMoon, HiOutlineSun } from 'react-icons/hi';
-import { Panel } from 'react-resizable-panels';
+import General from './General.jsx';
+import Profile from './Profile.jsx';
+import Preferences from './Preferences.jsx';
 
 export default function Settings({ isOpen, onClose, isDark, setIsDark }) {
-
+  
+  const [activeTab, setActiveTab] = useState('General');
+  
   useEffect(() => {
     if (!isOpen) return undefined;
 
@@ -17,6 +21,11 @@ export default function Settings({ isOpen, onClose, isDark, setIsDark }) {
     window.addEventListener('keydown', handleKeyDown);
     return () => window.removeEventListener('keydown', handleKeyDown);
   }, [isOpen, onClose]);
+
+  const tabs = ['General', 'Profile', 'Preferences'];
+  const isActive = (tab) => activeTab === tab;
+
+
 
   if (!isOpen) return null;
 
@@ -34,8 +43,8 @@ export default function Settings({ isOpen, onClose, isDark, setIsDark }) {
         aria-modal="true"
         aria-label="Settings"
       >
-        <div className="relative mb-4 flex items-center justify-center">
-          <h2 className="text-lg font-semibold">Settings</h2>
+        <div className="relative mb-4  justify-center">
+          {/*<h2 className="text-lg font-semibold">Settings</h2>*/}
           <button
             className="absolute right-0 flex h-8 w-8 items-center justify-center rounded-full hover:bg-hover dark:hover:bg-hover-dark"
             onClick={onClose}
@@ -56,19 +65,42 @@ className="flex h-[calc(100%-3rem)] gap-4"
   borderRadius: "0.5rem",
   width: "10rem",
   backgroundColor: isDark ? "var(--color-secondary-dark)" : "var(--color-secondary)",
+  gap: "0.5rem",
+  padding: "0.5rem",
 } 
 }>
 
 <button
-className="w-full rounded-md px-3 py-2 text-left hover:bg-hover dark:hover:bg-hover-dark">
+className={`w-full rounded-md px-3 py-2 text-left transition-colors ${
+  isActive('General')
+    ? 'bg-hover dark:bg-hover-dark font-medium'
+    : 'hover:bg-hover dark:hover:bg-hover-dark'
+}`}
+onClick={() => setActiveTab('General')}
+aria-pressed={isActive('General')}
+>
   General
 </button>
 
-<button className="w-full rounded-md px-3 py-2 text-left hover:bg-hover dark:hover:bg-hover-dark">
+<button className={`w-full rounded-md px-3 py-2 text-left transition-colors ${
+  isActive('Profile')
+    ? 'bg-hover dark:bg-hover-dark font-medium'
+    : 'hover:bg-hover dark:hover:bg-hover-dark'
+}`}
+onClick={() => setActiveTab('Profile')}
+aria-pressed={isActive('Profile')}
+>
   Profile
 </button>
 
-<button className="w-full rounded-md px-3 py-2 text-left hover:bg-hover dark:hover:bg-hover-dark">
+<button className={`w-full rounded-md px-3 py-2 text-left transition-colors ${
+  isActive('Preferences')
+    ? 'bg-hover dark:bg-hover-dark font-medium'
+    : 'hover:bg-hover dark:hover:bg-hover-dark'
+}`}
+onClick={() => setActiveTab('Preferences')}
+aria-pressed={isActive('Preferences')}
+>
   Preferences
 </button>
 
@@ -76,27 +108,17 @@ className="w-full rounded-md px-3 py-2 text-left hover:bg-hover dark:hover:bg-ho
 
 </aside>
   
-               <div className="justify left space-y-2 text-center">
-          <button
-            onClick={() => setIsDark(!isDark)}
-            className="flex w-full items-center justify-center gap-2 rounded-md px-3 py-2 hover:bg-hover dark:hover:bg-hover-dark"
-          >
-            {isDark ? <HiOutlineSun /> : <HiOutlineMoon />}
-            {isDark ? 'Light Mode' : 'Dark Mode'}
-          </button>
-          <p className="text-sm text-icon dark:text-icon-dark">
-            More to come
-          </p>
+               <div className="space-y-2 text-center">
+
+              {activeTab === 'General' && <General isDark={isDark} setIsDark={setIsDark} />}
+              {activeTab === 'Profile' && <Profile />}
+              {activeTab === 'Preferences' && <Preferences />}
+
         </div>
       </div>
+      </div>
     </div>  
-    
-
-</div>
-
-
-
-
+  
     
   );
 }
